Add exception rules for the unknown compiler component and fix the forge test models failures (#1217)
1. Analysed latest ran [model analysis
pipeline](https://github.com/tenstorrent/tt-forge-fe/actions/runs/13226767422)
artifacts (i.e
[script-outputs](https://github.com/tenstorrent/tt-forge-fe/actions/runs/13226767422/artifacts/2576589360)
) and found out there are some failed model test cases which includes
skipped model variants and fixed the model test cases failures in
`forge/test/models/pytorch` directory path
Note: In the model analysis pipeline, we are running all the models in
the `forge/test/models/pytorch` directory path which includes skipped
model test cases too.

2. Added new exception rules for the UNKNOWN Compiler component failures
present in the
[PR](https://github.com/tenstorrent/tt-forge-fe/pull/1210) for updating
model_analysis_docs with the help of
[unique-ops-logs](https://github.com/tenstorrent/tt-forge-fe/actions/runs/13226767422/artifacts/2576589743)
artifacts

+++ forge/forge/tvm_unique_op_generation.py
@@ -569,7 +569,7 @@ def extract_and_generate_unique_ops_tests(
         named_buffers_file_name = os.path.join(module_directory, str(current_module_name) + "_named_buffers.pt")
         torch.save(named_buffers, named_buffers_file_name)
     else:
-        if os.path.exists(param_file_name):
+        if param_file_name is not None and os.path.exists(param_file_name):
             os.remove(param_file_name)
             param_file_name = None
